websocket-service updated and fully worked, агай бая финалда стать продавцом иштебей атканы websocket-service'ти eureka'га кошпой калып api-gateway аны таппай калыптыр. Чатыбыз дагы ушул websocket-service аркылуу иштейт

+++ product-service/src/main/java/kg/buyers/productservice/dto/ReviewDTO.java
@@ -7,7 +7,7 @@ public class ReviewDTO {
     private Product productId;
     private String reviewerId;//user_id
     private String orderId;
-    private Short grade;
+    private Integer grade;
     private String options;
     private String text;
     private String images;
